Set up pages, some styling, app structure, and more
Our app structure is laid out with templates, components, and static folders. gitignore updated to weed out more stuff we don't need in the repo. requirements.txt is created to make installing dependencies easy. Readme instructions updated for setting up the app in Windows. Several pages and their routes are created. Base.html is created to serve as a base document for the rest of the app. The login form works, but needs additional validation and session info.