Created edit-classes form, conditional nav links, logout route, added style breakpoints
The edit classes form is on a separate page from the create-account form for reusability. I tried doing it without AJAX since I didn't know anything about it, but I had to learn some stuff. The form currently does not do anything with the data, since that will need to be handled in the edit-classes route. A stub is present to work off of. Conditional nav links show if the user is logged in, but they need additional styling. The logout route works. Breakpoints have been added to the style sheet.
fixes #8
fixes #15
fixes #20