Logout button alignment fix
I just had to add to the btn-danger (class="navbar-nav" and style="margin: 0; padding: 0") and then in the style.css I made it so that the button margin and border match the href links so that they all sit at the same height horizontally. Austin requested this quick fix lol.