Fixing all the screwiness with working across branches
It isn't the cleanest or best way to do it. I'm grappling with figuring out what goes where. Fortunately, all seems to work. Please test the app and figure out if anything is still broken. Please make sure you guys have the latest code at all times and are working off the main branch. Commit to the main branch. Working off other branches with different versions of the code is not doing us any favors this close to the deadline.